r600g/sb: Don't require array declarations for TGSI_FILE_SYSTEM_VALUE
Although gl_SampleMaskIn is declared as an array in GLSL, it is
effectively a 32 bit mask on all hardware supported by mesa, so the
array indexing is ignored (Thanks Glenn Kennard for the explanation).

Add a comment that the assert is not made superfluos by the else branch.

Corrects: piglit spec@arb_gpu_shader5@execution@samplemaskin-indirect for
debug builds (it already passed in release).