mesa/st/glsl_to_tgsi: Correct debug output for indirect access
For arrays print the array ID, and with indirect access also print the
reladdr* registers. The reladdr* registers are always used in the
printout, even though the actual code may use an address register.

Specifically, a sequence involving src.reladdr = TEMP[2] and src.index=10
that emits the address register loading instruction will be printed like:

  MOV ADDR[0].x, TEMP[2].xxxx
  MOV TEMP[3], ARRAY(2)[TEMP[2].xxxx + 10]

The reason for this is, that there is currently no indication in the src
register on whether the address instruction was or must be emitted.