r600/sb: bail out if prepare_alu_group() doesn't find a proper scheduling
It is possible that the optimizing ends up in an infinite loop in schedule_alu,
because prepare_alu_group() does not find a proper scheduling. This can be
deducted from pending.count() > 0 and not getting smaller.

This patch works around this problem by signalling this failure so that the
optimizers bails out and the un-optimized shader is used.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=103142