configure+mesa/st: check for -std=c++11 support and enable mesa/st/test accordingly
Add a check that tests whether the c++ compiler supports c++11, either by default,
by adding the compiler flag -std=c++11, or by adding a compiler flag that the user
has specified via the environment variable CXX11_CXXFLAGS.

The test only does a very shallow check of c++11 support, i.e. it tests whether
the define  __cplusplus >= 201103L to confirm language support by the compiler,
and if checks whether the header <tuple> is available to test the availability
of the c++11 standard library.

A make file conditional HAVE_STD_CXX11 is provided that is used in this patch to
to enable the tests in mesa/st if C++11 support is available.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=102665