R600: shader CF_OP_VTX also doesn't use the EOP bit.
Although the EOP bit is documented for the vertex fetch clause, it is not
properly interpreted. As a result the piglit
  spec/arb_tessellation_shader/execution/trivial-tess-gs_no-gs-inputs
creates a TESS_EVAL shader that does not have an EOP clause, which might
result in a GPU lockup.

This patch forces an additional NOP group like it is already done for other
CF_OP types.