mesa/st/glsl_to_tgsi: add array life range evaluation into tracking code
This adds the array live range tracking harvessting the already implemented
live range tracking for temporary registers. With this patch the array live
range tracking is not exposed, and since some tests that track reladdr access
require an array live range tracker to be availabe, so one is added as a fake.