mesa/st/glsl_to_tgsi: Add method to collect some statistics
When mesa is compiled in debug mode then this adds the possibility
to print out some statistics about the translated shaders to a file.

The functionality is enabled by setting the environment variable

   GLSL_TO_TGSI_PRINT_STATS

to the file name where the statistics should be collected. The file is
opened in append mode so that statistics from various runs will be
accumulated.