r600/sb: Don't assert when the a constraint is the same like the old
When a phi-node is updated ra_constraint::update_values() fires an assert
when the constraint is already set.

In cases where the new value for the constraint is the same like the old
value this seems to be unnecessary. In fact, the tests

  dEQP-GLES3.functional.shaders.metamorphic.bubblesort_flag.variant_1
  dEQP-GLES3.functional.shaders.metamorphic.bubblesort_flag.variant_2

that trigger this assertion because after the gvn optimizer run they have
phi nodes

  * phi     R18.x.10,	     R14.x.4, R14.x.4
  * phi     AR.9,            R14.x.4, R14.x.4

pass in release mode (and with the assertion disabled).

It may be that having a phi-node depening twice on the same constraint constitues a
bug in sb